refactor(csg-viewer): * removed redundant code * moving out reusable bits * etc
