--- packages/ui/csg-viewer/src/bound-utils/index.js
@@ -1,2 +1,0 @@
-export {default as computeBounds} from './computeBounds'
-export {default as isObjectOutsideBounds} from './isObjectOutsideBounds'
